Review of index.md
Mostly some rephrasing here and there as suggestions. Nothing major. The only thing I would urge you to make consistent is the use of Cumundi/they (mostly at the beginning of the article) vs. us/our (and in some cases my - no problem to write in the I-form but when discussing Cumundi code etc. it's clearer to use our). Some comments here and there that are tagged with "@Ringo".